feat: reorganize and rename contracts
