+++ contracts/base/Versionable.sol
@@ -2,7 +2,7 @@
 
 pragma solidity ^0.8.0;
 
-import "./interfaces/IVersionable.sol";
+import "../interfaces/IVersionable.sol";
 
 /**
  * @title Versionable contract
