+++ contracts/interfaces/ICreditLineTypes.sol
@@ -2,16 +2,10 @@
 
 pragma solidity 0.8.24;
 
-import { ICreditLine } from "./core/ICreditLine.sol";
-
-/// @title ICreditLineConfigurable interface
+/// @title ICreditLineTypes interface
 /// @author CloudWalk Inc. (See https://cloudwalk.io)
-/// @dev Defines the configurable credit line contract functions and events.
-interface ICreditLineConfigurable is ICreditLine {
-    // -------------------------------------------- //
-    //  Structs and enums                           //
-    // -------------------------------------------- //
-
+/// @dev Defines types that are used in the credit line contract.
+interface ICreditLineTypes {
     /// @dev Defines the available borrow policies.
     ///
     /// Possible values:
@@ -80,54 +74,4 @@ interface ICreditLineConfigurable is ICreditLine {
         uint64 totalClosedLoanAmount;
         // uint96 __reserved; // Reserved for future use until the end of the storage slot.
     }
-
-    // -------------------------------------------- //
-    //  Events                                      //
-    // -------------------------------------------- //
-
-    /// @dev Emitted when the credit line is configured.
-    /// @param creditLine The address of the current credit line.
-    event CreditLineConfigured(address indexed creditLine);
-
-    /// @dev Emitted when a borrower is configured.
-    /// @param creditLine The address of the current credit line.
-    /// @param borrower The address of the borrower being configured.
-    event BorrowerConfigured(address indexed creditLine, address indexed borrower);
-
-    // -------------------------------------------- //
-    //  Functions                                   //
-    // -------------------------------------------- //
-
-    /// @dev Updates the credit line configuration.
-    /// @param config The structure containing the credit line configuration.
-    function configureCreditLine(CreditLineConfig memory config) external;
-
-    /// @dev Configures a specific borrower.
-    /// @param borrower The address of the borrower to configure.
-    /// @param config The struct containing the borrower configuration.
-    function configureBorrower(address borrower, BorrowerConfig memory config) external;
-
-    /// @dev Configures multiple borrowers at once.
-    /// @param borrowers The addresses of the borrowers to configure.
-    /// @param configs The array containing the borrower configurations.
-    function configureBorrowers(address[] memory borrowers, BorrowerConfig[] memory configs) external;
-
-    /// @dev Retrieves the configuration of a borrower.
-    /// @param borrower The address of the borrower to check.
-    /// @return The structure containing the borrower configuration.
-    function getBorrowerConfiguration(address borrower) external view returns (BorrowerConfig memory);
-
-    /// @dev Retrieves the state of a borrower.
-    /// @param borrower The address of the borrower to check.
-    /// @return The structure containing the borrower state.
-    function getBorrowerState(address borrower) external view returns (BorrowerState memory);
-
-    /// @dev Retrieves the credit line configuration.
-    /// @return The structure containing the credit line configuration.
-    function creditLineConfiguration() external view returns (CreditLineConfig memory);
-
-    /// @dev Checks whether an account is an admin.
-    /// @param account The address of the account to check.
-    /// @return True if the account is configured as an admin.
-    function isAdmin(address account) external view returns (bool);
 }
