+++ contracts/libraries/SafeCast.sol
