+++ contracts/mocks/AccessControlExtUpgradeableMock.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.20;
 
 import { UUPSUpgradeable } from "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
 
-import { AccessControlExtUpgradeable } from "../common/AccessControlExtUpgradeable.sol";
+import { AccessControlExtUpgradeable } from "../base/AccessControlExtUpgradeable.sol";
 
 /**
  * @title AccessControlExtUpgradeableMock contract
