+++ contracts/mocks/LiquidityPoolMock.sol
@@ -2,15 +2,15 @@
 
 pragma solidity 0.8.24;
 
-import { Error } from "../common/libraries/Error.sol";
-import { ILiquidityPool } from "../common/interfaces/core/ILiquidityPool.sol";
 import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import { ILendingMarket } from "../common/interfaces/core/ILendingMarket.sol";
+
+import { Error } from "../libraries/Error.sol";
+import { ILendingMarket } from "../interfaces/ILendingMarket.sol";
 
 /// @title LiquidityPoolMock contract
 /// @author CloudWalk Inc. (See https://cloudwalk.io)
 /// @dev Mock of the `LiquidityPool` contract used for testing.
-contract LiquidityPoolMock is ILiquidityPool {
+contract LiquidityPoolMock {
     // -------------------------------------------- //
     //  Events                                      //
     // -------------------------------------------- //
@@ -24,8 +24,6 @@ contract LiquidityPoolMock is ILiquidityPool {
     // -------------------------------------------- //
     //  Storage variables                           //
     // -------------------------------------------- //
-
-    address private _tokenAddress;
 
     bool private _onBeforeLoanTakenResult;
 
@@ -54,18 +52,6 @@ contract LiquidityPoolMock is ILiquidityPool {
         return _onAfterLoanRevocationResult;
     }
 
-    function market() external pure returns (address) {
-        revert Error.NotImplemented();
-    }
-
-    function lender() external pure returns (address) {
-        revert Error.NotImplemented();
-    }
-
-    function token() external view returns (address) {
-        return _tokenAddress;
-    }
-
     function addonTreasury() external view returns (address) {
         return _addonTreasury;
     }
@@ -73,10 +59,6 @@ contract LiquidityPoolMock is ILiquidityPool {
     // -------------------------------------------- //
     //  Mock functions                              //
     // -------------------------------------------- //
-
-    function mockTokenAddress(address tokenAddress) external {
-        _tokenAddress = tokenAddress;
-    }
 
     function approveMarket(address _market, address token_) external {
         IERC20(token_).approve(_market, type(uint56).max);
