--- test/liquidity-pools/LiquidityPoolAccountableUUPS.test.ts
@@ -1,75 +1,0 @@
-import { ethers, upgrades } from "hardhat";
-import { expect } from "chai";
-import { Contract, ContractFactory } from "ethers";
-import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
-import { checkContractUupsUpgrading, connect } from "../../test-utils/eth";
-import { setUpFixture } from "../../test-utils/common";
-
-const ERROR_NAME_ACCESS_CONTROL_UNAUTHORIZED = "AccessControlUnauthorizedAccount";
-const ERROR_NAME_IMPLEMENTATION_ADDRESS_INVALID = "ImplementationAddressInvalid";
-
-const OWNER_ROLE = ethers.id("OWNER_ROLE");
-
-describe("Contract 'LiquidityPoolAccountableUUPS'", async () => {
-  let liquidityPoolFactory: ContractFactory;
-
-  let deployer: HardhatEthersSigner;
-  let market: HardhatEthersSigner;
-  let token: HardhatEthersSigner;
-  let attacker: HardhatEthersSigner;
-
-  before(async () => {
-    [deployer, market, token, attacker] = await ethers.getSigners();
-
-    liquidityPoolFactory = await ethers.getContractFactory("LiquidityPoolAccountableUUPS");
-    liquidityPoolFactory = liquidityPoolFactory.connect(deployer); // Explicitly specifying the deployer account
-  });
-
-  async function deployLiquidityPool(): Promise<{ liquidityPool: Contract }> {
-    let liquidityPool = await upgrades.deployProxy(
-      liquidityPoolFactory,
-      [deployer.address, market.address, token.address],
-      { kind: "uups" }
-    );
-
-    liquidityPool = connect(liquidityPool, deployer); // Explicitly specifying the initial account
-
-    return {
-      liquidityPool
-    };
-  }
-
-  describe("Function 'initialize()'", async () => {
-    it("Configures the contract as expected", async () => {
-      const { liquidityPool } = await setUpFixture(deployLiquidityPool);
-
-      expect(await liquidityPool.hasRole(OWNER_ROLE, deployer.address)).to.eq(true);
-      expect(await liquidityPool.market()).to.eq(market.address);
-    });
-  });
-
-  describe("Function 'upgradeToAndCall()'", async () => {
-    it("Executes as expected", async () => {
-      const { liquidityPool } = await setUpFixture(deployLiquidityPool);
-      await checkContractUupsUpgrading(liquidityPool, liquidityPoolFactory);
-    });
-
-    it("Is reverted if the caller is not the owner", async () => {
-      const { liquidityPool } = await setUpFixture(deployLiquidityPool);
-
-      await expect(connect(liquidityPool, attacker).upgradeToAndCall(liquidityPool, "0x"))
-        .to.be.revertedWithCustomError(liquidityPool, ERROR_NAME_ACCESS_CONTROL_UNAUTHORIZED)
-        .withArgs(attacker.address, OWNER_ROLE);
-    });
-
-    it("Is reverted if the provided implementation address is not a liquidity pool contract", async () => {
-      const { liquidityPool } = await setUpFixture(deployLiquidityPool);
-      const mockContractFactory = await ethers.getContractFactory("UUPSExtUpgradeableMock");
-      const mockContract = await mockContractFactory.deploy() as Contract;
-      await mockContract.waitForDeployment();
-
-      await expect(liquidityPool.upgradeToAndCall(mockContract, "0x"))
-        .to.be.revertedWithCustomError(liquidityPool, ERROR_NAME_IMPLEMENTATION_ADDRESS_INVALID);
-    });
-  });
-});
